requirements.txt
numpy==1.21.4
Pillow==8.4.0
torch==1.10.0+cu102
torchaudio==0.10.0+cu102
torchvision==0.11.1+cu102
typing-extensions==4.0.1
